hub-core/data-processor/requirements.txt
faust-streaming==0.10.14
mode-streaming==0.3.5
aiokafka==0.8.1
kafka-python==2.0.2
redis==5.0.1
asyncpg==0.29.0
python-dotenv==1.0.0
numpy
pandas
yarl==1.9.4
multidict==6.0.5
